backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "google-genai>=1.0",
    "groq>=0.9",
    "httpx>=0.27",
    "langfuse>=3.0",
    "openai>=1.51",
    "pillow>=10.3",
    "pydantic>=2.7",
    "pyjwt>=2.8",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "supabase>=2.5",
    "uvicorn[standard]>=0.30",
]
